backend/requirements.txt
fastapi==0.95.1 
uvicorn==0.22.0 
SQLAlchemy==2.0.13
psycopg2-binary==2.9.6
python-dotenv==1.0.0
alembic==1.11.1